Allowing one to use rails-like nested attributes when calling toJSON.
By passing { nested: true } to the toJSON method like this:

	model.toJSON({ nested: true })

It will return the keys of all relations with a _attributes, and also the deleted models in the collection that were once persisted, with a { _destroy: true } attribute.

This conforms with Rails nested attributes support when using accept_nested_attributes_for in a ActiveModel model.