Moving models that were marked to be destroyed to the deletedModels collection
For rails-like nested attributes after validating data on server it can return a serialized model with a `{ _destroy: true }` attribute. This means that the model is marked to be destroyed, but since there was a validation error it was not actually destroyed. By moving the model to the deletedModels collection we keep the state of the nested collection synchronized with the server.